requirements.txt
streamlit
google-generativeai
python-dotenv
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
PyPDF2
python-docx
Pillow
langchain
langchain-text-splitters
torchvision
torchaudio
pytesseract
Flask
PyMuPDF
selenium